fix : Footer 수정
